pyproject.toml
[project]
name = "helper-tools"
version = "2.4.0"
description = "Tools for automating/enhancing NixOS (and other) system tasks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich>=14.2.0",
    "pydantic-settings>=2.10.1",
]

[project.scripts]
py_motd = "py_motd.__main__:main"
podman_sops = "podman_sops.__main__:main"
py_backup = "py_backup.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podman_sops", "src/py_backup", "src/py_motd"]
